// src/store/layoutNavbarType.d.ts
export declare namespace layoutTabbarType {
  interface setNavbar {
    title?: string
    'left-text'?: string
    leftText?: string
    'right-text'?: string
    rightText?: string
    'left-arrow'?: boolean
    leftArrow?: boolean
    border?: boolean
    fixed?: boolean
    placeholder?: boolean
    'z-index'?: number | string
    zIndex?: number | string
    'safe-area-inset-top'?: boolean
    safeAreaInsetTop?: boolean
  }
}